move tests so they actually get executed
